twitter/backend/requirements.txt
Django==3.2.6
graphene-django==2.15.0
psycopg2-binary==2.9.1
gunicorn==19.9.0

django-allow-cidr==0.3.1
django-cors-headers==3.7.0
django-storages==1.11.1
django-ses==1.0.3
django-ipware==3.0.2
django-ratelimit==3.0.1
environs==9.3.0
Faker==7.0.1
flake8==3.8.4
graphene-django==2.15.0
django-graphql-jwt==0.3.0
django-graphql-auth